move tests for tracker to `test_inference` add tests for post_processing
