+++ tests/test_models.py
@@ -342,65 +342,3 @@ def test_tracking_transformer():
 
     assert asso_preds[0].size() == (num_detected * num_frames,) * 2
     assert embedding.size() == (num_detected * num_frames, 1, feats)
-
-
-def test_tracker():
-    feats = 512
-    num_frames = 5
-    num_detected = 2
-    img_shape = (1, 128, 128)
-    test_frame = 2
-    instances = []
-
-    for i in range(num_frames):
-        instances.append(
-            {
-                "frame_id": torch.tensor(i),
-                "img_shape": torch.tensor(img_shape),
-                "num_detected": torch.tensor([num_detected]),
-                "crops": torch.rand(size=(num_detected, 1, 64, 64)),
-                "bboxes": torch.rand(size=(num_detected, 4)),
-                "gt_track_ids": torch.arange(num_detected),
-                "pred_track_ids": torch.tensor([-1] * num_detected),
-            }
-        )
-
-    embedding_meta = {
-        "embedding_type": "fixed_pos",
-        "kwargs": {"temperature": num_detected, "scale": num_frames, "normalize": True},
-    }
-
-    tracking_transformer = GlobalTrackingTransformer(
-        encoder_model="resnet18",
-        encoder_cfg={"weights": "ResNet18_Weights.DEFAULT"},
-        d_model=feats,
-        num_encoder_layers=1,
-        num_decoder_layers=1,
-        dim_feedforward=feats,
-        feature_dim_attn_head=feats,
-        embedding_meta=embedding_meta,
-        return_embedding=False,
-    )
-    print("testing with no post processing")
-    tracking_cfg = {
-        "overlap_thresh": 0.01,
-        "decay_time": None,
-        "iou": None,
-        "max_center_dist": None,
-    }
-
-    tracker = Tracker(model=tracking_transformer, **tracking_cfg)
-
-    instances_pred = tracker(instances)
-
-    asso_equals = (
-        instances_pred[test_frame]["decay_time_traj_score"].to_numpy()
-        == instances_pred[test_frame]["with_iou_traj_score"].to_numpy()
-    ).all()
-    assert asso_equals
-
-    asso_equals = (
-        instances_pred[test_frame]["with_iou_traj_score"].to_numpy()
-        == instances_pred[test_frame]["max_center_dist_traj_score"].to_numpy()
-    ).all()
-    assert asso_equals
